files added for jwt authentication

+++ product-management-system/src/main/java/com/example/productmanagementsystem/service/UserService.java
@@ -6,10 +6,10 @@ import com.example.productmanagementsystem.entity.User;
 
 public interface UserService {
     User getUseByEmail(String email);
-    User getUsrById(Long id);
+    User getUsrById(Integer id);
     User saveUser(UserDto user);
-    User updateUser(Long id, User user);
-    User deleteUser(Long id);
+    User updateUser(Integer id, User user);
+    User deleteUser(Integer id);
 
     User authenticate(AuthReqDto requestDto);
 }
